requirements.txt
aiogram==3.20.0
asyncpg==0.30.0
redis==6.1.0
python-dotenv==1.1.0
gTTS==2.5.4
asyncio==3.4.3 # Python 3.12+ uchun
